Issue 12024: Made part of text document shared content immutable, refactored database mapping classes
git-svn-id: https://svn.imcode.com/imcode/imcms/trunk@30120 bd66a97b-2aff-0310-9095-89ca5cabf5a6

+++ src/main/java/com/imcode/imcms/mapping/orm/ContentLoop.java
@@ -87,7 +87,7 @@ public class ContentLoop implements Serializable, Cloneable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private volatile Long id;
+    private volatile Integer id;
 
     private volatile Integer no;
 
@@ -110,7 +110,7 @@ public class ContentLoop implements Serializable, Cloneable {
     protected ContentLoop() {
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
